Separate able for easier control of value interactions
