+++ src/components/text-input/TextInput.root.tsx
@@ -2,24 +2,32 @@ import styled, { css } from 'styled-components';
 import { FlattenSimpleInterpolation, ThemedStyledProps } from 'styled-components/ts3.6';
 
 import { applyShape } from '../../foundations/shape';
-import { setAlpha, blendColors } from '../../shared';
-import { applyPadding } from '../../utilities/layout';
+import { setAlpha, blendColors, Utils } from '../../shared';
+import { applyMargin, applyPadding } from '../../utilities/layout';
 import { TTheme } from '../../utilities/theme';
 import Spacing from '../../utilities/spacing';
+import { applyTextStyles } from '../text';
 
-import { TEXT_INPUT_VALUES_MAPPING } from './TextInput.constants';
+import { TEXT_INPUT_VALUES_MAPPING, LABEL_POSITIONS } from './TextInput.constants';
 import { PTextInputRoot } from './TextInput.props';
 
 const TextInputRoot = styled.div<PTextInputRoot>(
     ({
-        theme: { palette, action, text, background },
+        theme: { palette, action, text, background, animation },
         hasError,
         disabled,
         active,
+        animatedLabel,
+        value,
         width,
         size,
+        leadingIcon,
+        onClear,
         backgroundColor = background.default,
     }: ThemedStyledProps<PTextInputRoot, TTheme>): FlattenSimpleInterpolation => {
+        const hasValue = Utils.isString(value) && value.length > 0;
+        const isClearable = Utils.isFunction(onClear);
+
         const colors: Record<string, string> = {
             active: hasError ? palette.alert.main : palette.primary.main,
             text: text.primary,
@@ -53,6 +61,13 @@ const TextInputRoot = styled.div<PTextInputRoot>(
                       border-color: ${colors.active};
                       background-color: ${setAlpha(colors.active, 0.16)};
                   }
+
+                  input:focus + span {
+                      background-color: ${backgroundColor};
+                      ${animatedLabel
+                          ? `transform: translate(${LABEL_POSITIONS[size]}) scale(0.7); color: ${colors.active};`
+                          : 'transform: scale(0); opacity: 0;'};
+                  }
                   &:focus-within {
                       border: 2px solid ${colors.active};
                   }
@@ -67,7 +82,6 @@ const TextInputRoot = styled.div<PTextInputRoot>(
             display: flex;
             align-items: center;
             background-color: ${colors.background};
-
             ${applyShape({
                 radius: 4,
                 width: width === 'full' ? '100%' : `${width}px`,
@@ -79,6 +93,34 @@ const TextInputRoot = styled.div<PTextInputRoot>(
                     horizontal: TEXT_INPUT_VALUES_MAPPING[size].spacing,
                 })
             )};
+
+            span {
+                position: absolute;
+                white-space: nowrap;
+                background-color: ${hasValue ? backgroundColor : 'transparent'};
+                color: ${colors.text};
+                opacity: 1;
+                transform: translate(${hasValue ? LABEL_POSITIONS[size] : '0, 0'})
+                    scale(${hasValue ? '0.7' : '1'});
+                ${applyShape({ radius: 4 })};
+                ${applyPadding(Spacing.symmetric({ vertical: 25, horizontal: 75 }))};
+                ${applyMargin(
+                    Spacing.only(
+                        'left',
+                        leadingIcon === 'none' ? 0 : TEXT_INPUT_VALUES_MAPPING[size].labelMargin
+                    )
+                )};
+                ${applyTextStyles({
+                    size: TEXT_INPUT_VALUES_MAPPING[size].labelSize,
+                    inheritLineHeight: true,
+                })};
+
+                transition: all ${animation.fastest} ease-in-out;
+            }
+
+            .clear {
+                cursor: ${isClearable ? 'pointer' : 'none'};
+            }
         `;
     }
 );
